Move completer into extra module
