+++ src/lib.rs
@@ -114,6 +114,8 @@ pub use self::error::{Result, Error, ErrorKind};
 
 #[macro_use] mod macros;
 pub mod busybox;
+#[cfg(feature="readline")]
+pub mod completer;
 #[cfg(feature="network")]
 pub mod crypto;
 pub mod ctrl;
