+++ src/shell.rs
@@ -2,10 +2,6 @@
 
 use busybox;
 use clap;
-#[cfg(feature="readline")]
-use rustyline;
-#[cfg(feature="readline")]
-use rustyline::completion::Completer;
 
 use Error;
 use ErrorKind;
@@ -64,39 +60,6 @@ impl From<NativeCommand> for Command {
 impl From<ForeignCommand> for Command {
     fn from(cmd: ForeignCommand) -> Command {
         Command::Foreign(cmd)
-    }
-}
-
-
-pub struct CmdCompleter(Arc<Mutex<Toolbox>>);
-
-impl CmdCompleter {
-    #[inline]
-    pub fn new(toolbox: Arc<Mutex<Toolbox>>) -> CmdCompleter {
-        CmdCompleter(toolbox)
-    }
-
-    #[inline]
-    #[cfg(feature="readline")]
-    fn commands(&self) -> Vec<String> {
-        self.0.lock().unwrap().keys()
-    }
-}
-
-#[cfg(feature="readline")]
-impl Completer for CmdCompleter {
-    #[inline]
-    fn complete(&self, line: &str, pos: usize) -> rustyline::Result<(usize, Vec<String>)> {
-        if line.contains(' ') || line.len() != pos {
-            return Ok((0, vec![]));
-        }
-
-        let results: Vec<String> = self.commands().iter()
-            .filter(|x| x.starts_with(line))
-            .map(|x| x.clone() + " ")
-            .collect();
-
-        Ok((0, results))
     }
 }
